redid routes for the dashboard
