+++ controllers/api/postRoutes.js
@@ -2,7 +2,37 @@ const router = require("express").Router();
 const { Post } = require("../../models");
 const withAuth = require("../../utils/auth");
 
-router.post('/', withAuth, async (req, res) => {
+// display all posts on dashboard
+router.get("/", withAuth, async (req, res) => {
+  try {
+    // Get all projects and JOIN with user data
+    const postData = await Post.findAll({
+      where: {
+        user_id: req.session.user_id,
+      },
+    });
+
+    // Serialize data so the template can read it
+    const posts = postData.map((post) => post.get({ plain: true }));
+
+    // Pass serialized data and session flag into template
+    res.render("dashboard", {
+      posts,
+      logged_in: req.session.logged_in,
+    });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+// route for creating a new blog post
+router.get("/postCreate", withAuth, async (req, res) => {
+  res.render("postCreate", {
+    logged_in: req.session.logged_in,
+  });
+});
+
+router.post("/postCreate", withAuth, async (req, res) => {
   try {
     const newPost = await Post.create({
       ...req.body,
@@ -15,6 +45,37 @@ router.post('/', withAuth, async (req, res) => {
   }
 });
 
+// route for viewing a sinlge blog post
+router.get("/postUpdate/:id", withAuth, async (req, res) => {
+  try {
+    const postData = await Post.findByPk(req.params.id, {});
+
+    const post = postData.get({ plain: true });
+
+    res.render("postUpdate", {
+      ...post,
+      logged_in: req.session.logged_in,
+    });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+// route for creating a new blog post
+router.post("/", withAuth, async (req, res) => {
+  try {
+    const newPost = await Post.create({
+      ...req.body,
+      user_id: req.session.user_id,
+    });
+
+    res.status(200).json(newPost);
+  } catch (err) {
+    res.status(400).json(err);
+  }
+});
+
+// route for deleting a blog post
 router.delete("/:id", withAuth, async (req, res) => {
   try {
     const postData = await Post.destroy({
@@ -35,13 +96,14 @@ router.delete("/:id", withAuth, async (req, res) => {
   }
 });
 
-// create put route
-router.put("/:id", withAuth, async (req, res) => {
+// route for updating an existing blog post
+// router.put("/:id", withAuth, async (req, res) => {
+router.put("/postUpdate/:id", async (req, res) => {
   try {
     const postData = await Post.update(req.body, {
       where: {
         id: req.params.id,
-        user_id: req.session.user_id,
+        // user_id: req.session.user_id,
       },
     });
