--- controllers/dashboardRoutes.js
@@ -1,87 +1,0 @@
-// show all loggedin users posts
-// create a post
-// edit and delete a post
-//
-
-const router = require("express").Router();
-const { Post, User } = require("../models");
-const withAuth = require("../utils/auth");
-
-router.get("/", withAuth, async (req, res) => {
-  try {
-    // Get all projects and JOIN with user data
-    const postData = await Post.findAll({
-      where: {
-        user_id: req.session.user_id,
-      },
-    });
-
-    // Serialize data so the template can read it
-    const posts = postData.map((post) => post.get({ plain: true }));
-
-    // Pass serialized data and session flag into template
-    res.render("dashboard", {
-      posts,
-      logged_in: req.session.logged_in,
-    });
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-router.get("/postCreate", withAuth, async (req, res) => {
-  res.render("postCreate", {
-    logged_in: req.session.logged_in,
-  });
-});
-
-router.post("/postCreate", withAuth, async (req, res) => {
-  try {
-    const newPost = await Post.create({
-      ...req.body,
-      user_id: req.session.user_id,
-    });
-
-    res.status(200).json(newPost);
-  } catch (err) {
-    res.status(400).json(err);
-  }
-});
-
-router.get("/postUpdate/:id", withAuth, async (req, res) => {
-  try {
-    const postData = await Post.findByPk(req.params.id, {});
-
-    const post = postData.get({ plain: true });
-
-    res.render("postUpdate", {
-      ...post,
-      logged_in: req.session.logged_in,
-    });
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-router.put("/:id", withAuth, async (req, res) => {
-  try {
-    const postData = await Post.update(req.body, {
-      where: {
-        id: req.params.id,
-        user_id: req.session.user_id,
-      },
-    });
-
-    if (!postData) {
-      res.status(404).json({ message: "No post found with this id!" });
-      return;
-    }
-
-    res.status(200).json(postData);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-
-module.exports = router;
