+++ public/js/postCreate.js
@@ -13,7 +13,7 @@ const postFormHandler = async (event) => {
 
     if (response.ok) {
       // If successful, redirect the browser to the dashboard page
-      document.location.replace("/dashboard");
+      document.location.replace("/api/posts");
     } else {
       alert(response.statusText);
     }
