+++ views/layouts/main.handlebars
@@ -19,11 +19,11 @@
         {{#if logged_in}}
         <!-- Conditionally render login or logout links -->
         <a href="/">homepage</a> |
-        <a href="/dashboard">dashboard</a> |
+        <a href="/api/posts">dashboard</a> |
         <button class="no-button" id="logout">logout</button>
         {{else}}
         <a href="/">homepage</a> |
-        <a href="/dashboard">dashboard</a> |
+        <a href="/api/posts">dashboard</a> |
         <a href="/login">login</a>
         {{/if}}
       </nav>
